requirements.txt
streamlit
google.generativeai
pypdf
opencv-python
python-docx
pillow
